added loads of notices and conditionals to skip unnecessary parts

+++ apply.sh
@@ -1,8 +1,12 @@
 #!/usr/bin/sh
 
 #custom executables (mostly rerouting and fixed args)
-cp .bin_replacements ~/ -r
-chmod -w ~/.bin_replacements -R #for security
+if [ ! -f ~/.bin_replacements ]; then
+	cp .bin_replacements ~/ -r
+	chmod -w ~/.bin_replacements -R #for security
+else
+	echo "skipped bin replacements"
+fi
 #making sure all the folders exist
 mkdir ~/.config/kitty ~/.config/qtile ~/.config/sway ~/.config/waybar \
 ~/.config/qt5ct/qss ~/.config/Kvantum/ArcDark# ~/.config/gtk-2.0 \
@@ -45,6 +49,8 @@ if [ ! -f ~/.local/lib/deadbeef/discord_presence.so ]; then
 	mv ~/.local/lib/deadbeef/plugins/discord_presence.so ~/.local/lib/deadbeef/discord_presence.so
 	rm ~/.local/lib/deadbeef/plugins/ -d
 	rm /tmp/ddb_discord_presence.zip
+else
+	echo "skipped deadbeef discord presence plugin"
 fi
 if [ ! -f ~/.local/lib/deadbeef/opus.so ]; then
 	mkdir /data/diyfs/ddb_opus_plugin -p
@@ -53,6 +59,8 @@ if [ ! -f ~/.local/lib/deadbeef/opus.so ]; then
 	make
 	mv ./opus.so ~/.local/lib/deadbeef/
 	cd ~/Scripts/dotfiles
+else
+	echo "skipped deadbeef opus plugin"
 fi
 if [ ! -f ~/.local/lib/deadbeef/seekbar.so ]; then
 	mkdir /data/diyfs/ddb_waveform_seekbar -p
@@ -62,19 +70,37 @@ if [ ! -f ~/.local/lib/deadbeef/seekbar.so ]; then
 	mv ./gtk2/ddb_misc_waveform_GTK2.so ~/.local/lib/deadbeef/ddb_misc_waveform_GTK2.so
 	mv ./gtk3/ddb_misc_waveform_GTK3.so ~/.local/lib/deadbeef/ddb_misc_waveform_GTK3.so
 	cd ~/Scripts/dotfiles
+else
+	echo "skipped deadbeef waveform seekbar plugin"
 fi
 #Music
-mkdir ~/Music/RYTD -p
-git clone https://github.com/RiedleroD/RYTD.git ~/Music/RYTD
-cp playlist.rpl ~/Music/default.rpl
-cp .rytdconf ~/Music/RYTD/.rytdconf
-python3 /home/riedler/Music/RYTD/rytd.py
+if [ ! -f ~/Music/RYTD ]; then
+	mkdir ~/Music/RYTD -p
+	git clone https://github.com/RiedleroD/RYTD.git ~/Music/RYTD
+	cp playlist.rpl ~/Music/default.rpl
+	cp .rytdconf ~/Music/RYTD/.rytdconf
+	python3 /home/riedler/Music/RYTD/rytd.py
+else
+	echo "skipped RYTD"
+fi
 #package compilation config
-doas cp makepkg.conf /etc/makepkg.conf
+if cmp -s -- makepkg.conf /etc/makepkg.conf; then
+	doas cp makepkg.conf /etc/makepkg.conf
+else
+	echo "skipped makepkg.conf"
+fi
 #wallpapers
 mkdir /home/riedler/Pictures -p
-curl https://i.imgur.com/2kyvrtg.png > /home/riedler/Pictures/wallpaper.png
-curl https://i.imgur.com/e2rqx1O.jpg > /home/riedler/Pictures/wallpaper.jpg
+if [ ! -f ~/Pictures/wallpaper.png ]; then
+	curl https://i.imgur.com/2kyvrtg.png > ~/Pictures/wallpaper.png
+else
+	echo "wallpaper.png skipped"
+fi
+if [ ! -f ~/Pictures/wallpaper.jpg ]; then
+	curl https://i.imgur.com/e2rqx1O.jpg > ~/Pictures/wallpaper.jpg
+else
+	echo "wallpaper.jpg skipped"
+fi
 
 #notice :)
 echo "Make sure you installed the dependencies!"
